trip-chat-service/pyproject.toml
[project]
name = "chat"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ag-ui-adk>=0.5.2",
    "fastapi[standard]>=0.123.10",
    "google-adk[a2a,otel-gcp]>=1.28.0",
    "grpcio>=1.75.1",
    "litellm>=1.80.5",
    "mcp-weather-server>=0.6.1",
    "mem0ai>=0.1.115",
    "nacos-sdk-python>=3.0.3",
    "openinference-instrumentation-google-adk>=0.1.8",
    "openinference-instrumentation-litellm>=0.1.29",
    "opentelemetry-distro[otlp]>=0.59b0",
    "pydantic-settings>=2.11.0",
    "pymongo>=4.15.3",
    "tripsphere>=0.1.0",
]

[build-system]
requires = ["uv_build>=0.10.0,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pyright>=1.1.406",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-mock>=3.15.1",
    "types-grpcio>=1.0.0.20251009",
    "types-protobuf>=6.32.1.20251210",
]

[tool.uv.sources]
tripsphere = { workspace = true }

[tool.uv.workspace]
members = ["libs/tripsphere"]

[tool.ruff]
target-version = "py312"
extend-exclude = ["libs/tripsphere"]

[tool.ruff.lint]
select = ["E", "F", "B", "I"]

[tool.mypy]
strict = true
python_version = "3.12"

[tool.pyright]
typeCheckingMode = "strict"
include = ["src", "tests"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-q"
